Onedirectional checkins
Squashed commit of the following:

commit 767bb5129089a513ab29aaef71a9caf105645217
Author: Jonas Vacek <1302278+jvacek@users.noreply.github.com>
Date:   Mon Apr 20 12:54:29 2026 +0200

    enable check in button for signed out users

commit ec738109754a3d9683aab250647996a0f343f49c
Author: Jonas Vacek <1302278+jvacek@users.noreply.github.com>
Date:   Mon Apr 20 12:52:36 2026 +0200

    improve view when disabled

commit 0a9be044c55b4582251d909782c85bc56c61b773
Author: Jonas Vacek <1302278+jvacek@users.noreply.github.com>
Date:   Mon Apr 20 12:36:04 2026 +0200

    test checkin permissions

commit 0f3fb5e6a4825d191df4ad92d156ef08494868d2
Author: Jonas Vacek <1302278+jvacek@users.noreply.github.com>
Date:   Mon Apr 20 12:34:30 2026 +0200

    one-driectional checkins

+++ backend/api/views.py
@@ -162,6 +162,12 @@ class CheckInViewSet(ListModelMixin, CreateModelMixin, UpdateModelMixin, Destroy
         if unit.admin_only_checkin and not (self.request.user.is_superuser or self.request.user.is_staff):
             msg = "This unit can only be checked in by admins."
             raise PermissionDenied(msg)
+        if not unit.can_user_check_in(self.request.user):
+            msg = (
+                "You can't check in here \u2014 once someone else takes the lighter, "
+                "its journey moves on. You can still follow along by subscribing."
+            )
+            raise PermissionDenied(msg)
         serializer.save(created_by=self.request.user, unit=unit)
         unit.subscribers.add(self.request.user)
         cache.delete_many([unit_distance_cache_key(unit.identifier), STATS_CACHE_KEY, GLOBE_PINS_CACHE_KEY])
